requirements.txt
absl-py==2.3.1
audioop-lts==0.2.2
pydub==0.25.1
piper-tts==1.3.0
srt==3.5.3
